force current to be negative instead of int overflow

+++ lunabot_embedded/firmware/teensy_main/lib/lunabot_drivers/StepperLib.cpp
@@ -117,7 +117,7 @@ int StepperMotor::read_torque() {
 
 // amps
 float StepperMotor::read_current() {
-  int x = read_register(Addrs.Read_Current);
+  int x = (int16_t)read_register(Addrs.Read_Current);
   if (x == -1) {
     return -1;
   }
